refactor(auth/pres): decouple login endpoint from app layer
- Replace LoginCommand with LoginRequest DTO as endpoint param
- Create Contracts.Login namespace to store DTO by logical use
- Use AddDays() instead of Add(TimeSpan.FromDays()) for clarity
