+++ services/auth/src/Auth.Presentation/Contracts/Login/LoginRequest.cs
@@ -1,0 +1,6 @@
+namespace Auth.Presentation.Contracts.Login;
+
+public sealed record LoginRequest(
+    string Email,
+    string Password
+);
